+++ server/src/main/java/org/opensearch/index/store/DirectoryFileTransferTracker.java
@@ -8,6 +8,7 @@
 
 package org.opensearch.index.store;
 
+import org.apache.lucene.store.Directory;
 import org.opensearch.common.util.MovingAverage;
 import org.opensearch.core.common.io.stream.StreamInput;
 import org.opensearch.core.common.io.stream.StreamOutput;
@@ -15,128 +16,150 @@ import org.opensearch.core.common.io.stream.Writeable;
 
 import java.io.IOException;
 import java.util.Objects;
+import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicReference;
 
 /**
- * Tracks the amount of bytes transferred between two {@link org.apache.lucene.store.Directory} instances
+ * Tracks the amount of bytes transferred between two {@link Directory} instances
  *
  * @opensearch.internal
  */
 public class DirectoryFileTransferTracker {
     /**
-     * Cumulative size of files (in bytes) attempted to be transferred over from the source {@link org.apache.lucene.store.Directory}
+     * Cumulative size of files (in bytes) attempted to be transferred over from the source {@link Directory}
      */
-    private volatile long transferredBytesStarted;
+    private final AtomicLong transferredBytesStarted = new AtomicLong();
 
     /**
-     * Cumulative size of files (in bytes) successfully transferred over from the source {@link org.apache.lucene.store.Directory}
+     * Cumulative size of files (in bytes) successfully transferred over from the source {@link Directory}
      */
-    private volatile long transferredBytesFailed;
+    private final AtomicLong transferredBytesFailed = new AtomicLong();
 
     /**
-     * Cumulative size of files (in bytes) failed in transfer over from the source {@link org.apache.lucene.store.Directory}
+     * Cumulative size of files (in bytes) failed in transfer over from the source {@link Directory}
      */
-    private volatile long transferredBytesSucceeded;
+    private final AtomicLong transferredBytesSucceeded = new AtomicLong();
 
     /**
-     * Time in milliseconds for the last successful transfer from the source {@link org.apache.lucene.store.Directory}
+     * Time in milliseconds for the last successful transfer from the source {@link Directory}
      */
-    private volatile long lastTransferTimestampMs;
+    private final AtomicLong lastTransferTimestampMs = new AtomicLong();
 
     /**
-     * Provides moving average over the last N total size in bytes of files transferred from the source {@link org.apache.lucene.store.Directory}.
+     * Cumulative time in milliseconds spent in successful transfers from the source {@link Directory}
+     */
+    private final AtomicLong totalTransferTimeInMs = new AtomicLong();
+
+    /**
+     * Provides moving average over the last N total size in bytes of files transferred from the source {@link Directory}.
      * N is window size
      */
-    private volatile MovingAverage transferredBytesMovingAverageReference;
+    private final AtomicReference<MovingAverage> transferredBytesMovingAverageReference;
 
-    private volatile long lastSuccessfulTransferInBytes;
+    private final AtomicLong lastSuccessfulTransferInBytes = new AtomicLong();
 
     /**
-     * Provides moving average over the last N transfer speed (in bytes/s) of segment files transferred from the source {@link org.apache.lucene.store.Directory}.
+     * Provides moving average over the last N transfer speed (in bytes/s) of segment files transferred from the source {@link Directory}.
      * N is window size
      */
-    private volatile MovingAverage transferredBytesPerSecMovingAverageReference;
+    private final AtomicReference<MovingAverage> transferredBytesPerSecMovingAverageReference;
 
     private final int DIRECTORY_FILES_TRANSFER_DEFAULT_WINDOW_SIZE = 20;
 
+    // Getters and Setters, all are visible for testing
     public long getTransferredBytesStarted() {
-        return transferredBytesStarted;
+        return transferredBytesStarted.get();
     }
 
     public void addTransferredBytesStarted(long size) {
-        transferredBytesStarted += size;
+        transferredBytesStarted.getAndAdd(size);
     }
 
     public long getTransferredBytesFailed() {
-        return transferredBytesFailed;
+        return transferredBytesFailed.get();
     }
 
-    public void addTransferredBytesFailed(long size) {
-        transferredBytesFailed += size;
+    public void addTransferredBytesFailed(long size, long startTimeInMs) {
+        transferredBytesFailed.getAndAdd(size);
+        addTotalTransferTimeInMs(Math.max(1, System.currentTimeMillis() - startTimeInMs));
     }
 
     public long getTransferredBytesSucceeded() {
-        return transferredBytesSucceeded;
+        return transferredBytesSucceeded.get();
     }
 
     public void addTransferredBytesSucceeded(long size, long startTimeInMs) {
-        transferredBytesSucceeded += size;
-        updateLastSuccessfulTransferSize(size);
+        transferredBytesSucceeded.getAndAdd(size);
+        updateSuccessfulTransferSize(size);
         long currentTimeInMs = System.currentTimeMillis();
         updateLastTransferTimestampMs(currentTimeInMs);
         long timeTakenInMS = Math.max(1, currentTimeInMs - startTimeInMs);
+        addTotalTransferTimeInMs(timeTakenInMS);
         addTransferredBytesPerSec((size * 1_000L) / timeTakenInMS);
     }
 
     public boolean isTransferredBytesPerSecAverageReady() {
-        return transferredBytesPerSecMovingAverageReference.isReady();
+        return transferredBytesPerSecMovingAverageReference.get().isReady();
     }
 
     public double getTransferredBytesPerSecAverage() {
-        return transferredBytesPerSecMovingAverageReference.getAverage();
+        return transferredBytesPerSecMovingAverageReference.get().getAverage();
     }
 
-    // Visible for testing
     public void addTransferredBytesPerSec(long bytesPerSec) {
-        this.transferredBytesPerSecMovingAverageReference.record(bytesPerSec);
+        this.transferredBytesPerSecMovingAverageReference.get().record(bytesPerSec);
     }
 
     public boolean isTransferredBytesAverageReady() {
-        return transferredBytesMovingAverageReference.isReady();
+        return transferredBytesMovingAverageReference.get().isReady();
     }
 
     public double getTransferredBytesAverage() {
-        return transferredBytesMovingAverageReference.getAverage();
+        return transferredBytesMovingAverageReference.get().getAverage();
     }
 
-    // Visible for testing
-    public void updateLastSuccessfulTransferSize(long size) {
-        lastSuccessfulTransferInBytes = size;
-        this.transferredBytesMovingAverageReference.record(size);
+    public void updateLastSuccessfulTransferInBytes(long size) {
+        lastSuccessfulTransferInBytes.set(size);
+    }
+
+    public void updateSuccessfulTransferSize(long size) {
+        updateLastSuccessfulTransferInBytes(size);
+        this.transferredBytesMovingAverageReference.get().record(size);
     }
 
     public long getLastTransferTimestampMs() {
-        return lastTransferTimestampMs;
+        return lastTransferTimestampMs.get();
     }
 
-    // Visible for testing
     public void updateLastTransferTimestampMs(long downloadTimestampInMs) {
-        this.lastTransferTimestampMs = downloadTimestampInMs;
+        this.lastTransferTimestampMs.set(downloadTimestampInMs);
+    }
+
+    public void addTotalTransferTimeInMs(long totalTransferTimeInMs) {
+        this.totalTransferTimeInMs.addAndGet(totalTransferTimeInMs);
+    }
+
+    public long getTotalTransferTimeInMs() {
+        return totalTransferTimeInMs.get();
     }
 
     public DirectoryFileTransferTracker() {
-        transferredBytesMovingAverageReference = new MovingAverage(DIRECTORY_FILES_TRANSFER_DEFAULT_WINDOW_SIZE);
-        transferredBytesPerSecMovingAverageReference = new MovingAverage(DIRECTORY_FILES_TRANSFER_DEFAULT_WINDOW_SIZE);
+        transferredBytesMovingAverageReference = new AtomicReference<>(new MovingAverage(DIRECTORY_FILES_TRANSFER_DEFAULT_WINDOW_SIZE));
+        transferredBytesPerSecMovingAverageReference = new AtomicReference<>(
+            new MovingAverage(DIRECTORY_FILES_TRANSFER_DEFAULT_WINDOW_SIZE)
+        );
     }
 
     public DirectoryFileTransferTracker.Stats stats() {
         return new Stats(
-            transferredBytesStarted,
-            transferredBytesFailed,
-            transferredBytesSucceeded,
-            lastTransferTimestampMs,
-            transferredBytesMovingAverageReference.getAverage(),
-            lastSuccessfulTransferInBytes,
-            transferredBytesPerSecMovingAverageReference.getAverage()
+            transferredBytesStarted.get(),
+            transferredBytesFailed.get(),
+            transferredBytesSucceeded.get(),
+            lastTransferTimestampMs.get(),
+            totalTransferTimeInMs.get(),
+            transferredBytesMovingAverageReference.get().getAverage(),
+            lastSuccessfulTransferInBytes.get(),
+            transferredBytesPerSecMovingAverageReference.get().getAverage()
         );
     }
 
@@ -150,6 +173,7 @@ public class DirectoryFileTransferTracker {
         public final long transferredBytesFailed;
         public final long transferredBytesSucceeded;
         public final long lastTransferTimestampMs;
+        public final long totalTransferTimeInMs;
         public final double transferredBytesMovingAverage;
         public final long lastSuccessfulTransferInBytes;
         public final double transferredBytesPerSecMovingAverage;
@@ -159,6 +183,7 @@ public class DirectoryFileTransferTracker {
             long transferredBytesFailed,
             long downloadBytesSucceeded,
             long lastTransferTimestampMs,
+            long totalTransferTimeInMs,
             double transferredBytesMovingAverage,
             long lastSuccessfulTransferInBytes,
             double transferredBytesPerSecMovingAverage
@@ -167,6 +192,7 @@ public class DirectoryFileTransferTracker {
             this.transferredBytesFailed = transferredBytesFailed;
             this.transferredBytesSucceeded = downloadBytesSucceeded;
             this.lastTransferTimestampMs = lastTransferTimestampMs;
+            this.totalTransferTimeInMs = totalTransferTimeInMs;
             this.transferredBytesMovingAverage = transferredBytesMovingAverage;
             this.lastSuccessfulTransferInBytes = lastSuccessfulTransferInBytes;
             this.transferredBytesPerSecMovingAverage = transferredBytesPerSecMovingAverage;
@@ -177,6 +203,7 @@ public class DirectoryFileTransferTracker {
             this.transferredBytesFailed = in.readLong();
             this.transferredBytesSucceeded = in.readLong();
             this.lastTransferTimestampMs = in.readLong();
+            this.totalTransferTimeInMs = in.readLong();
             this.transferredBytesMovingAverage = in.readDouble();
             this.lastSuccessfulTransferInBytes = in.readLong();
             this.transferredBytesPerSecMovingAverage = in.readDouble();
@@ -188,6 +215,7 @@ public class DirectoryFileTransferTracker {
             out.writeLong(transferredBytesFailed);
             out.writeLong(transferredBytesSucceeded);
             out.writeLong(lastTransferTimestampMs);
+            out.writeLong(totalTransferTimeInMs);
             out.writeDouble(transferredBytesMovingAverage);
             out.writeLong(lastSuccessfulTransferInBytes);
             out.writeDouble(transferredBytesPerSecMovingAverage);
@@ -203,6 +231,7 @@ public class DirectoryFileTransferTracker {
                 && transferredBytesFailed == stats.transferredBytesFailed
                 && transferredBytesSucceeded == stats.transferredBytesSucceeded
                 && lastTransferTimestampMs == stats.lastTransferTimestampMs
+                && totalTransferTimeInMs == stats.totalTransferTimeInMs
                 && Double.compare(stats.transferredBytesMovingAverage, transferredBytesMovingAverage) == 0
                 && lastSuccessfulTransferInBytes == stats.lastSuccessfulTransferInBytes
                 && Double.compare(stats.transferredBytesPerSecMovingAverage, transferredBytesPerSecMovingAverage) == 0;
@@ -215,6 +244,7 @@ public class DirectoryFileTransferTracker {
                 transferredBytesFailed,
                 transferredBytesSucceeded,
                 lastTransferTimestampMs,
+                totalTransferTimeInMs,
                 transferredBytesMovingAverage,
                 lastSuccessfulTransferInBytes,
                 transferredBytesPerSecMovingAverage