+++ server/src/test/java/org/opensearch/action/admin/cluster/node/stats/NodeStatsTests.java
@@ -459,6 +459,9 @@ public class NodeStatsTests extends OpenSearchTestCase {
                     assertEquals(remoteSegmentStats.getUploadBytesFailed(), deserializedRemoteSegmentStats.getUploadBytesFailed());
                     assertEquals(remoteSegmentStats.getMaxRefreshTimeLag(), deserializedRemoteSegmentStats.getMaxRefreshTimeLag());
                     assertEquals(remoteSegmentStats.getMaxRefreshBytesLag(), deserializedRemoteSegmentStats.getMaxRefreshBytesLag());
+                    assertEquals(remoteSegmentStats.getTotalRefreshBytesLag(), deserializedRemoteSegmentStats.getTotalRefreshBytesLag());
+                    assertEquals(remoteSegmentStats.getTotalUploadTime(), deserializedRemoteSegmentStats.getTotalUploadTime());
+                    assertEquals(remoteSegmentStats.getTotalDownloadTime(), deserializedRemoteSegmentStats.getTotalDownloadTime());
                 }
             }
         }
@@ -789,8 +792,11 @@ public class NodeStatsTests extends OpenSearchTestCase {
             remoteSegmentStats.addDownloadBytesStarted(10L);
             remoteSegmentStats.addDownloadBytesSucceeded(10L);
             remoteSegmentStats.addDownloadBytesFailed(1L);
-            remoteSegmentStats.setMaxRefreshBytesLag(5L);
+            remoteSegmentStats.addTotalRefreshBytesLag(5L);
+            remoteSegmentStats.addMaxRefreshBytesLag(2L);
             remoteSegmentStats.setMaxRefreshTimeLag(2L);
+            remoteSegmentStats.addTotalUploadTime(20L);
+            remoteSegmentStats.addTotalDownloadTime(20L);
         }
         return indicesStats;
     }